requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
Django==4.0
django-admin==2.0.2
django-apscheduler==0.6.2
django-environ==0.10.0
django-excel-response2==3.0.5
django-filter==23.1
django-six==1.0.5
django-tinymce==3.6.1
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tzdata==2023.2
tzlocal==4.3
xlwt==1.3.0
